requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
cvxpy==1.3.1
ecos==2.0.12
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysql-connector-python-rf==2.2.2
numpy==1.24.3
osqp==0.6.2.post9
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
qdldl==0.1.7
scipy==1.10.1
scs==3.2.3
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
